refactor: user module
