+++ server/apps/immich/src/api-v1/album/album.module.ts
@@ -6,7 +6,6 @@ import { AlbumEntity, AssetAlbumEntity, UserAlbumEntity } from '@app/database';
 import { AlbumRepository, IAlbumRepository } from './album-repository';
 import { DownloadModule } from '../../modules/download/download.module';
 import { AssetModule } from '../asset/asset.module';
-import { UserModule } from '../user/user.module';
 import { ShareModule } from '../share/share.module';
 
 const ALBUM_REPOSITORY_PROVIDER = {
@@ -18,7 +17,6 @@ const ALBUM_REPOSITORY_PROVIDER = {
   imports: [
     TypeOrmModule.forFeature([AlbumEntity, AssetAlbumEntity, UserAlbumEntity]),
     DownloadModule,
-    UserModule,
     forwardRef(() => AssetModule),
     ShareModule,
   ],
