+++ server/apps/immich/src/api-v1/oauth/oauth.module.ts
@@ -1,12 +1,11 @@
 import { ImmichConfigModule } from '@app/immich-config';
 import { Module } from '@nestjs/common';
 import { ImmichJwtModule } from '../../modules/immich-jwt/immich-jwt.module';
-import { UserModule } from '../user/user.module';
 import { OAuthController } from './oauth.controller';
 import { OAuthService } from './oauth.service';
 
 @Module({
-  imports: [UserModule, ImmichJwtModule, ImmichConfigModule],
+  imports: [ImmichJwtModule, ImmichConfigModule],
   controllers: [OAuthController],
   providers: [OAuthService],
   exports: [OAuthService],
