+++ server/apps/immich/src/api-v1/oauth/oauth.service.ts
@@ -5,9 +5,7 @@ import { ClientMetadata, custom, generators, Issuer, UserinfoResponse } from 'op
 import { AuthUserDto } from '../../decorators/auth-user.decorator';
 import { ImmichJwtService } from '../../modules/immich-jwt/immich-jwt.service';
 import { LoginResponseDto } from '../auth/response-dto/login-response.dto';
-import { UserResponseDto } from '../user/response-dto/user-response.dto';
-import { IUserRepository } from '@app/domain';
-import { UserCore } from '../user/user.core';
+import { IUserRepository, UserResponseDto, UserCore } from '@app/domain';
 import { OAuthCallbackDto } from './dto/oauth-auth-code.dto';
 import { OAuthConfigDto } from './dto/oauth-config.dto';
 import { OAuthConfigResponseDto } from './response-dto/oauth-config-response.dto';
