+++ server/apps/immich/src/config/asset-upload.config.ts
@@ -51,7 +51,7 @@ function destination(req: Request, file: Express.Multer.File, cb: any) {
 
   const basePath = APP_UPLOAD_LOCATION;
   const sanitizedDeviceId = sanitize(String(req.body['deviceId']));
-  const originalUploadFolder = join(basePath, req.user.id, 'original', sanitizedDeviceId);
+  const originalUploadFolder = join(basePath, user.id, 'original', sanitizedDeviceId);
 
   if (!existsSync(originalUploadFolder)) {
     mkdirSync(originalUploadFolder, { recursive: true });
