+++ server/apps/immich/src/config/profile-image-upload.config.ts
@@ -6,6 +6,7 @@ import { existsSync, mkdirSync } from 'fs';
 import { diskStorage } from 'multer';
 import { extname } from 'path';
 import sanitize from 'sanitize-filename';
+import { AuthUserDto } from '../decorators/auth-user.decorator';
 import { patchFormData } from '../utils/path-form-data.util';
 
 export const profileImageUploadOption: MulterOptions = {
@@ -35,8 +36,10 @@ function destination(req: Request, file: Express.Multer.File, cb: any) {
     return cb(new UnauthorizedException());
   }
 
+  const user = req.user as AuthUserDto;
+
   const basePath = APP_UPLOAD_LOCATION;
-  const profileImageLocation = `${basePath}/${req.user.id}/profile`;
+  const profileImageLocation = `${basePath}/${user.id}/profile`;
 
   if (!existsSync(profileImageLocation)) {
     mkdirSync(profileImageLocation, { recursive: true });
