+++ server/apps/immich/src/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './user.controller';
