+++ server/apps/immich/src/decorators/auth-user.decorator.ts
@@ -1,14 +1,6 @@
+export { AuthUserDto } from '@app/domain';
+import { AuthUserDto } from '@app/domain';
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-// import { AuthUserDto } from './dto/auth-user.dto';
-
-export class AuthUserDto {
-  id!: string;
-  email!: string;
-  isAdmin!: boolean;
-  isPublicUser?: boolean;
-  sharedLinkId?: string;
-  isAllowUpload?: boolean;
-}
 
 export const GetAuthUser = createParamDecorator((data, ctx: ExecutionContext): AuthUserDto => {
   return ctx.switchToHttp().getRequest<{ user: AuthUserDto }>().user;
