+++ server/apps/immich/src/global.d.ts
@@ -1,8 +1,8 @@
-import { UserResponseDto } from './api-v1/user/response-dto/user-response.dto';
+import { AuthUserDto } from './decorators/auth-user.decorator';
 
 declare global {
   namespace Express {
     // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    interface User extends UserResponseDto {}
+    interface User extends AuthUserDto {}
   }
 }
