+++ server/libs/domain/src/auth/dto/auth-user.dto.ts
@@ -1,0 +1,8 @@
+export class AuthUserDto {
+  id!: string;
+  email!: string;
+  isAdmin!: boolean;
+  isPublicUser?: boolean;
+  sharedLinkId?: string;
+  isAllowUpload?: boolean;
+}
