+++ server/libs/domain/src/auth/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './auth-user.dto';
