+++ server/libs/domain/src/index.ts
@@ -1,2 +1,3 @@
+export * from './auth';
 export * from './domain.module';
 export * from './user';
