+++ server/libs/domain/src/user/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './create-profile-image.dto';
+export * from './create-user.dto';
+export * from './update-user.dto';
+export * from './user-count.dto';
