+++ server/libs/domain/src/user/index.ts
@@ -1,1 +1,5 @@
+export * from './dto';
+export * from './response-dto';
+export * from './user.core';
 export * from './user.repository';
+export * from './user.service';
