+++ server/libs/domain/src/user/response-dto/create-profile-image-response.dto.ts
