+++ server/libs/infra/src/index.ts
@@ -1,2 +1,1 @@
 export * from './infra.module';
-export * from './repository';
