+++ server/libs/infra/src/infra.module.ts
@@ -1,6 +1,6 @@
 import { UserEntity } from '@app/database';
 import { IUserRepository } from '@app/domain';
-import { Module, Provider } from '@nestjs/common';
+import { Global, Module, Provider } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from './repository';
 
@@ -9,6 +9,7 @@ const providers: Provider[] = [
   { provide: IUserRepository, useClass: UserRepository },
 ];
 
+@Global()
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity])],
   providers: [...providers],
