+++ server/libs/infra/src/repository/user.repository.ts
@@ -1,9 +1,10 @@
 import { UserEntity } from '@app/database';
 import { IUserRepository, UserListFilter } from '@app/domain';
-import { InternalServerErrorException } from '@nestjs/common';
+import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Not, Repository } from 'typeorm';
 
+@Injectable()
 export class UserRepository implements IUserRepository {
   constructor(
     @InjectRepository(UserEntity)
